+++ src/main/java/wooteco/subway/aop/exception/SubwayException.java
@@ -1,4 +1,4 @@
-package wooteco.subway;
+package wooteco.subway.aop.exception;
 
 public class SubwayException extends RuntimeException {
     private final String code;
